+++ src/zmq/zmqutil.h
@@ -5,6 +5,14 @@
 #ifndef BITCOIN_ZMQ_ZMQUTIL_H
 #define BITCOIN_ZMQ_ZMQUTIL_H
 
+#include <string>
+
 void zmqError(const char *str);
+
+/**
+ * Prefix for unix domain socket addresses (which are local filesystem paths)
+ * Used by libzmq, example "ipc:///root/path/to/file"
+ */
+const std::string ADDR_PREFIX_IPC = "ipc://";
 
 #endif // BITCOIN_ZMQ_ZMQUTIL_H