+++ doc/release-notes.md
@@ -5,3 +5,5 @@ Bitcoin ABC version 0.33.9 is now available from:
   <https://download.bitcoinabc.org/0.33.9/>
 
 This release includes the following features and fixes:
+- unix socket paths are now accepted for `-zmqpubrawblock` and `-zmqpubrawtx` with
+  the format `-zmqpubrawtx=unix:/path/to/file`